storage/requirements_provision.txt
azure-mgmt-resource
azure-mgmt-storage
azure-identity
